+++ src/siganalysis/__init__.py
@@ -3,6 +3,7 @@
 # Use of this source code is governed by a MIT-style license that
 # can be found in the LICENSE.txt file for the project.
 from .siganalysis import (
+    STFT_WINDOWS,
     calculate_peak_hold,
     freq_bin,
     hz2khz,
@@ -16,6 +17,7 @@ from .siganalysis import (
 )
 
 __all__ = [
+    "STFT_WINDOWS",
     "calculate_peak_hold",
     "freq_bin",
     "hz2khz",